requirements.txt
PyQt5-Qt5==5.15.16
ShopifyAPI==12.4.0
beautifulsoup4==4.12.3
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.1
playwright==1.41.2
python-whois==0.8.0
tldextract==5.1.1
openai==1.12.0
wikipedia==1.4.0
price-parser==0.4.0 